Closeness centrality implemented
Changes were made in the implementation of closeness centrality and in the tests, since we have changed the metric definition to something that makes sense to a directed graph, as is now dipected in the D2.1.

The page rank tests were @ignored, since I'm in doubt about the page rank definition: it's because we have a homogenous linear equations system, which always provide us a solution filled with zeros.