Tests refactored
Before we had the tests separeted by the graphs used as example.
It was done to minimize repeated code in tests.
But now we changed to separete the tests by funcionality,
since this makes easir to assess if a single functionality is working
(before we had to run many tests to find it out!).
This changes improves future refactorings evaluations.